requirements.txt
python-can[mf4]==4.5.0
colorama==0.4.6